Refactor scene management with store
