+++ package.json
@@ -18,6 +18,7 @@
     "typescript": "^4.3.5"
   },
   "dependencies": {
-    "p5": "^1.4.0"
+    "p5": "^1.4.0",
+    "zustand": "^3.5.7"
   }
 }
